+++ shell/browser/api/atom_api_system_preferences_mac.mm
@@ -218,14 +218,12 @@ int SystemPreferences::DoSubscribeNotification(
 
                 if (user_info) {
                   copied_callback.Run(
-                      base::SysNSStringToUTF8(notification.name),
-                      *user_info,
+                      base::SysNSStringToUTF8(notification.name), *user_info,
                       object);
                 } else {
                   copied_callback.Run(
                       base::SysNSStringToUTF8(notification.name),
-                      base::DictionaryValue(),
-                      object);
+                      base::DictionaryValue(), object);
                 }
               }];
   return request_id;
