+++ spec/api-browser-window-spec.js
@@ -9,7 +9,6 @@ const qs = require('querystring')
 const http = require('http')
 const { closeWindow } = require('./window-helpers')
 const { emittedOnce } = require('./events-helpers')
-const { createNetworkSandbox } = require('./network-helper')
 const { ipcRenderer, remote } = require('electron')
 const { app, ipcMain, BrowserWindow, BrowserView, protocol, session, screen, webContents } = remote
 
@@ -112,1384 +111,6 @@ describe('BrowserWindow module', () => {
   beforeEach(openTheWindow)
 
   afterEach(closeTheWindow)
-
-  describe('"webPreferences" option', () => {
-    afterEach(() => { ipcMain.removeAllListeners('answer') })
-
-    describe('"preload" option', () => {
-      const doesNotLeakSpec = (name, webPrefs) => {
-        it(name, async function () {
-          w.destroy()
-          w = new BrowserWindow({
-            webPreferences: {
-              ...webPrefs,
-              preload: path.resolve(fixtures, 'module', 'empty.js')
-            },
-            show: false
-          })
-          const leakResult = emittedOnce(ipcMain, 'leak-result')
-          w.loadFile(path.join(fixtures, 'api', 'no-leak.html'))
-          const [, result] = await leakResult
-          expect(result).to.have.property('require', 'undefined')
-          expect(result).to.have.property('exports', 'undefined')
-          expect(result).to.have.property('windowExports', 'undefined')
-          expect(result).to.have.property('windowPreload', 'undefined')
-          expect(result).to.have.property('windowRequire', 'undefined')
-        })
-      }
-      doesNotLeakSpec('does not leak require', {
-        nodeIntegration: false,
-        sandbox: false,
-        contextIsolation: false
-      })
-      doesNotLeakSpec('does not leak require when sandbox is enabled', {
-        nodeIntegration: false,
-        sandbox: true,
-        contextIsolation: false
-      })
-      doesNotLeakSpec('does not leak require when context isolation is enabled', {
-        nodeIntegration: false,
-        sandbox: false,
-        contextIsolation: true
-      })
-      doesNotLeakSpec('does not leak require when context isolation and sandbox are enabled', {
-        nodeIntegration: false,
-        sandbox: true,
-        contextIsolation: true
-      })
-
-      it('loads the script before other scripts in window', async () => {
-        const preload = path.join(fixtures, 'module', 'set-global.js')
-        w.destroy()
-        w = new BrowserWindow({
-          show: false,
-          webPreferences: {
-            nodeIntegration: true,
-            preload
-          }
-        })
-        const p = emittedOnce(ipcMain, 'answer')
-        w.loadFile(path.join(fixtures, 'api', 'preload.html'))
-        const [, test] = await p
-        expect(test).to.eql('preload')
-      })
-      it('can successfully delete the Buffer global', async () => {
-        const preload = path.join(fixtures, 'module', 'delete-buffer.js')
-        w.destroy()
-        w = new BrowserWindow({
-          show: false,
-          webPreferences: {
-            nodeIntegration: true,
-            preload
-          }
-        })
-        const p = emittedOnce(ipcMain, 'answer')
-        w.loadFile(path.join(fixtures, 'api', 'preload.html'))
-        const [, test] = await p
-        expect(test.toString()).to.eql('buffer')
-      })
-      it('has synchronous access to all eventual window APIs', async () => {
-        const preload = path.join(fixtures, 'module', 'access-blink-apis.js')
-        const w = await openTheWindow({
-          show: false,
-          webPreferences: {
-            nodeIntegration: true,
-            preload
-          }
-        })
-        const p = emittedOnce(ipcMain, 'answer')
-        w.loadFile(path.join(fixtures, 'api', 'preload.html'))
-        const [, test] = await p
-        expect(test).to.be.an('object')
-        expect(test.atPreload).to.be.an('array')
-        expect(test.atLoad).to.be.an('array')
-        expect(test.atPreload).to.deep.equal(test.atLoad, 'should have access to the same window APIs')
-      })
-    })
-
-    describe('session preload scripts', function () {
-      const preloads = [
-        path.join(fixtures, 'module', 'set-global-preload-1.js'),
-        path.join(fixtures, 'module', 'set-global-preload-2.js'),
-        path.relative(process.cwd(), path.join(fixtures, 'module', 'set-global-preload-3.js'))
-      ]
-      const defaultSession = session.defaultSession
-
-      beforeEach(() => {
-        expect(defaultSession.getPreloads()).to.deep.equal([])
-        defaultSession.setPreloads(preloads)
-      })
-      afterEach(() => {
-        defaultSession.setPreloads([])
-      })
-
-      it('can set multiple session preload script', function () {
-        expect(defaultSession.getPreloads()).to.deep.equal(preloads)
-      })
-
-      const generateSpecs = (description, sandbox) => {
-        describe(description, () => {
-          it('loads the script before other scripts in window including normal preloads', function (done) {
-            ipcMain.once('vars', function (event, preload1, preload2, preload3) {
-              expect(preload1).to.equal('preload-1')
-              expect(preload2).to.equal('preload-1-2')
-              expect(preload3).to.be.null()
-              done()
-            })
-            w.destroy()
-            w = new BrowserWindow({
-              show: false,
-              webPreferences: {
-                sandbox,
-                preload: path.join(fixtures, 'module', 'get-global-preload.js')
-              }
-            })
-            w.loadURL('about:blank')
-          })
-        })
-      }
-
-      generateSpecs('without sandbox', false)
-      generateSpecs('with sandbox', true)
-    })
-
-    describe('"additionalArguments" option', () => {
-      it('adds extra args to process.argv in the renderer process', (done) => {
-        const preload = path.join(fixtures, 'module', 'check-arguments.js')
-        ipcMain.once('answer', (event, argv) => {
-          expect(argv).to.include('--my-magic-arg')
-          done()
-        })
-        w.destroy()
-        w = new BrowserWindow({
-          show: false,
-          webPreferences: {
-            nodeIntegration: true,
-            preload,
-            additionalArguments: ['--my-magic-arg']
-          }
-        })
-        w.loadFile(path.join(fixtures, 'api', 'blank.html'))
-      })
-
-      it('adds extra value args to process.argv in the renderer process', (done) => {
-        const preload = path.join(fixtures, 'module', 'check-arguments.js')
-        ipcMain.once('answer', (event, argv) => {
-          expect(argv).to.include('--my-magic-arg=foo')
-          done()
-        })
-        w.destroy()
-        w = new BrowserWindow({
-          show: false,
-          webPreferences: {
-            nodeIntegration: true,
-            preload,
-            additionalArguments: ['--my-magic-arg=foo']
-          }
-        })
-        w.loadFile(path.join(fixtures, 'api', 'blank.html'))
-      })
-    })
-
-    describe('"node-integration" option', () => {
-      it('disables node integration by default', (done) => {
-        const preload = path.join(fixtures, 'module', 'send-later.js')
-        ipcMain.once('answer', (event, typeofProcess, typeofBuffer) => {
-          expect(typeofProcess).to.equal('undefined')
-          expect(typeofBuffer).to.equal('undefined')
-          done()
-        })
-        w.destroy()
-        w = new BrowserWindow({
-          show: false,
-          webPreferences: {
-            preload
-          }
-        })
-        w.loadFile(path.join(fixtures, 'api', 'blank.html'))
-      })
-    })
-
-    describe('"enableRemoteModule" option', () => {
-      const generateSpecs = (description, sandbox) => {
-        describe(description, () => {
-          const preload = path.join(fixtures, 'module', 'preload-remote.js')
-
-          it('enables the remote module by default', async () => {
-            const w = await openTheWindow({
-              show: false,
-              webPreferences: {
-                preload,
-                sandbox
-              }
-            })
-            const p = emittedOnce(ipcMain, 'remote')
-            w.loadFile(path.join(fixtures, 'api', 'blank.html'))
-            const [, remote] = await p
-            expect(remote).to.equal('object')
-          })
-
-          it('disables the remote module when false', async () => {
-            const w = await openTheWindow({
-              show: false,
-              webPreferences: {
-                preload,
-                sandbox,
-                enableRemoteModule: false
-              }
-            })
-            const p = emittedOnce(ipcMain, 'remote')
-            w.loadFile(path.join(fixtures, 'api', 'blank.html'))
-            const [, remote] = await p
-            expect(remote).to.equal('undefined')
-          })
-        })
-      }
-
-      generateSpecs('without sandbox', false)
-      generateSpecs('with sandbox', true)
-    })
-
-    describe('"sandbox" option', () => {
-      function waitForEvents (emitter, events, callback) {
-        let count = events.length
-        for (const event of events) {
-          emitter.once(event, () => {
-            if (!--count) callback()
-          })
-        }
-      }
-
-      const preload = path.join(fixtures, 'module', 'preload-sandbox.js')
-
-      it('exposes ipcRenderer to preload script', (done) => {
-        ipcMain.once('answer', function (event, test) {
-          expect(test).to.equal('preload')
-          done()
-        })
-        w.destroy()
-        w = new BrowserWindow({
-          show: false,
-          webPreferences: {
-            sandbox: true,
-            preload
-          }
-        })
-        w.loadFile(path.join(fixtures, 'api', 'preload.html'))
-      })
-
-      it('exposes ipcRenderer to preload script (path has special chars)', function (done) {
-        const preloadSpecialChars = path.join(fixtures, 'module', 'preload-sandboxæø åü.js')
-        ipcMain.once('answer', function (event, test) {
-          expect(test).to.equal('preload')
-          done()
-        })
-        w.destroy()
-        w = new BrowserWindow({
-          show: false,
-          webPreferences: {
-            sandbox: true,
-            preload: preloadSpecialChars
-          }
-        })
-        w.loadFile(path.join(fixtures, 'api', 'preload.html'))
-      })
-
-      it('exposes "loaded" event to preload script', function (done) {
-        w.destroy()
-        w = new BrowserWindow({
-          show: false,
-          webPreferences: {
-            sandbox: true,
-            preload
-          }
-        })
-        ipcMain.once('process-loaded', () => done())
-        w.loadURL('about:blank')
-      })
-
-      it('exposes "exit" event to preload script', function (done) {
-        w.destroy()
-        w = new BrowserWindow({
-          show: false,
-          webPreferences: {
-            sandbox: true,
-            preload
-          }
-        })
-        const htmlPath = path.join(fixtures, 'api', 'sandbox.html?exit-event')
-        const pageUrl = 'file://' + htmlPath
-        ipcMain.once('answer', function (event, url) {
-          let expectedUrl = pageUrl
-          if (process.platform === 'win32') {
-            expectedUrl = 'file:///' + htmlPath.replace(/\\/g, '/')
-          }
-          expect(url).to.equal(expectedUrl)
-          done()
-        })
-        w.loadURL(pageUrl)
-      })
-
-      it('should open windows in same domain with cross-scripting enabled', (done) => {
-        w.destroy()
-        w = new BrowserWindow({
-          show: false,
-          webPreferences: {
-            sandbox: true,
-            preload
-          }
-        })
-        ipcRenderer.send('set-web-preferences-on-next-new-window', w.webContents.id, 'preload', preload)
-        const htmlPath = path.join(fixtures, 'api', 'sandbox.html?window-open')
-        const pageUrl = 'file://' + htmlPath
-        w.webContents.once('new-window', (e, url, frameName, disposition, options) => {
-          let expectedUrl = pageUrl
-          if (process.platform === 'win32') {
-            expectedUrl = 'file:///' + htmlPath.replace(/\\/g, '/')
-          }
-          expect(url).to.equal(expectedUrl)
-          expect(frameName).to.equal('popup!')
-          expect(options.width).to.equal(500)
-          expect(options.height).to.equal(600)
-          ipcMain.once('answer', function (event, html) {
-            expect(html).to.equal('<h1>scripting from opener</h1>')
-            done()
-          })
-        })
-        w.loadURL(pageUrl)
-      })
-
-      it('should open windows in another domain with cross-scripting disabled', async () => {
-        const w = await openTheWindow({
-          show: false,
-          webPreferences: {
-            sandbox: true,
-            preload
-          }
-        })
-
-        ipcRenderer.send('set-web-preferences-on-next-new-window', w.webContents.id, 'preload', preload)
-        const openerWindowOpen = emittedOnce(ipcMain, 'opener-loaded')
-        w.loadFile(
-          path.join(fixtures, 'api', 'sandbox.html'),
-          { search: 'window-open-external' }
-        )
-
-        // Wait for a message from the main window saying that it's ready.
-        await openerWindowOpen
-
-        // Ask the opener to open a popup with window.opener.
-        const expectedPopupUrl = `${server.url}/cross-site` // Set in "sandbox.html".
-
-        const browserWindowCreated = emittedOnce(app, 'browser-window-created')
-        w.webContents.send('open-the-popup', expectedPopupUrl)
-
-        // The page is going to open a popup that it won't be able to close.
-        // We have to close it from here later.
-        // XXX(alexeykuzmin): It will leak if the test fails too soon.
-        const [, popupWindow] = await browserWindowCreated
-
-        // Ask the popup window for details.
-        const detailsAnswer = emittedOnce(ipcMain, 'child-loaded')
-        popupWindow.webContents.send('provide-details')
-        const [, openerIsNull, , locationHref] = await detailsAnswer
-        expect(openerIsNull).to.be.false('window.opener is null')
-        expect(locationHref).to.equal(expectedPopupUrl)
-
-        // Ask the page to access the popup.
-        const touchPopupResult = emittedOnce(ipcMain, 'answer')
-        w.webContents.send('touch-the-popup')
-        const [, popupAccessMessage] = await touchPopupResult
-
-        // Ask the popup to access the opener.
-        const touchOpenerResult = emittedOnce(ipcMain, 'answer')
-        popupWindow.webContents.send('touch-the-opener')
-        const [, openerAccessMessage] = await touchOpenerResult
-
-        // We don't need the popup anymore, and its parent page can't close it,
-        // so let's close it from here before we run any checks.
-        await closeWindow(popupWindow, { assertSingleWindow: false })
-
-        expect(popupAccessMessage).to.be.a('string',
-          `child's .document is accessible from its parent window`)
-        expect(popupAccessMessage).to.match(/^Blocked a frame with origin/)
-        expect(openerAccessMessage).to.be.a('string',
-          `opener .document is accessible from a popup window`)
-        expect(openerAccessMessage).to.match(/^Blocked a frame with origin/)
-      })
-
-      it('should inherit the sandbox setting in opened windows', (done) => {
-        w.destroy()
-        w = new BrowserWindow({
-          show: false,
-          webPreferences: {
-            sandbox: true
-          }
-        })
-
-        const preloadPath = path.join(fixtures, 'api', 'new-window-preload.js')
-        ipcRenderer.send('set-web-preferences-on-next-new-window', w.webContents.id, 'preload', preloadPath)
-        ipcMain.once('answer', (event, args) => {
-          expect(args).to.include('--enable-sandbox')
-          done()
-        })
-        w.loadFile(path.join(fixtures, 'api', 'new-window.html'))
-      })
-
-      it('should open windows with the options configured via new-window event listeners', (done) => {
-        w.destroy()
-        w = new BrowserWindow({
-          show: false,
-          webPreferences: {
-            sandbox: true
-          }
-        })
-
-        const preloadPath = path.join(fixtures, 'api', 'new-window-preload.js')
-        ipcRenderer.send('set-web-preferences-on-next-new-window', w.webContents.id, 'preload', preloadPath)
-        ipcRenderer.send('set-web-preferences-on-next-new-window', w.webContents.id, 'foo', 'bar')
-        ipcMain.once('answer', (event, args, webPreferences) => {
-          expect(webPreferences.foo).to.equal('bar')
-          done()
-        })
-        w.loadFile(path.join(fixtures, 'api', 'new-window.html'))
-      })
-
-      it('should set ipc event sender correctly', (done) => {
-        w.destroy()
-        w = new BrowserWindow({
-          show: false,
-          webPreferences: {
-            sandbox: true,
-            preload
-          }
-        })
-        ipcRenderer.send('set-web-preferences-on-next-new-window', w.webContents.id, 'preload', preload)
-
-        let childWc
-        w.webContents.once('new-window', (e, url, frameName, disposition, options) => {
-          childWc = options.webContents
-          expect(w.webContents).to.not.equal(childWc)
-        })
-        ipcMain.once('parent-ready', function (event) {
-          expect(w.webContents).to.equal(event.sender)
-          event.sender.send('verified')
-        })
-        ipcMain.once('child-ready', function (event) {
-          expect(childWc).to.be.an('object')
-          expect(childWc).to.equal(event.sender)
-          event.sender.send('verified')
-        })
-        waitForEvents(ipcMain, [
-          'parent-answer',
-          'child-answer'
-        ], done)
-        w.loadFile(path.join(fixtures, 'api', 'sandbox.html'), { search: 'verify-ipc-sender' })
-      })
-
-      describe('event handling', () => {
-        it('works for window events', (done) => {
-          waitForEvents(w, [
-            'page-title-updated'
-          ], done)
-          w.loadFile(path.join(fixtures, 'api', 'sandbox.html'), { search: 'window-events' })
-        })
-
-        it('works for stop events', (done) => {
-          waitForEvents(w.webContents, [
-            'did-navigate',
-            'did-fail-load',
-            'did-stop-loading'
-          ], done)
-          w.loadFile(path.join(fixtures, 'api', 'sandbox.html'), { search: 'webcontents-stop' })
-        })
-
-        it('works for web contents events', (done) => {
-          waitForEvents(w.webContents, [
-            'did-finish-load',
-            'did-frame-finish-load',
-            'did-navigate-in-page',
-            'will-navigate',
-            'did-start-loading',
-            'did-stop-loading',
-            'did-frame-finish-load',
-            'dom-ready'
-          ], done)
-          w.loadFile(path.join(fixtures, 'api', 'sandbox.html'), { search: 'webcontents-events' })
-        })
-      })
-
-      it('supports calling preventDefault on new-window events', (done) => {
-        w.destroy()
-        w = new BrowserWindow({
-          show: false,
-          webPreferences: {
-            sandbox: true
-          }
-        })
-        const initialWebContents = webContents.getAllWebContents().map((i) => i.id)
-        ipcRenderer.send('prevent-next-new-window', w.webContents.id)
-        w.webContents.once('new-window', () => {
-          // We need to give it some time so the windows get properly disposed (at least on OSX).
-          setTimeout(() => {
-            const currentWebContents = webContents.getAllWebContents().map((i) => i.id)
-            expect(currentWebContents).to.deep.equal(initialWebContents)
-            done()
-          }, 100)
-        })
-        w.loadFile(path.join(fixtures, 'pages', 'window-open.html'))
-      })
-
-      // see #9387
-      it('properly manages remote object references after page reload', (done) => {
-        w.destroy()
-        w = new BrowserWindow({
-          show: false,
-          webPreferences: {
-            preload,
-            sandbox: true
-          }
-        })
-        w.loadFile(path.join(fixtures, 'api', 'sandbox.html'), { search: 'reload-remote' })
-
-        ipcMain.on('get-remote-module-path', (event) => {
-          event.returnValue = path.join(fixtures, 'module', 'hello.js')
-        })
-
-        let reload = false
-        ipcMain.on('reloaded', (event) => {
-          event.returnValue = reload
-          reload = !reload
-        })
-
-        ipcMain.once('reload', (event) => {
-          event.sender.reload()
-        })
-
-        ipcMain.once('answer', (event, arg) => {
-          ipcMain.removeAllListeners('reloaded')
-          ipcMain.removeAllListeners('get-remote-module-path')
-          expect(arg).to.equal('hi')
-          done()
-        })
-      })
-
-      it('properly manages remote object references after page reload in child window', (done) => {
-        w.destroy()
-        w = new BrowserWindow({
-          show: false,
-          webPreferences: {
-            preload,
-            sandbox: true
-          }
-        })
-        ipcRenderer.send('set-web-preferences-on-next-new-window', w.webContents.id, 'preload', preload)
-
-        w.loadFile(path.join(fixtures, 'api', 'sandbox.html'), { search: 'reload-remote-child' })
-
-        ipcMain.on('get-remote-module-path', (event) => {
-          event.returnValue = path.join(fixtures, 'module', 'hello-child.js')
-        })
-
-        let reload = false
-        ipcMain.on('reloaded', (event) => {
-          event.returnValue = reload
-          reload = !reload
-        })
-
-        ipcMain.once('reload', (event) => {
-          event.sender.reload()
-        })
-
-        ipcMain.once('answer', (event, arg) => {
-          ipcMain.removeAllListeners('reloaded')
-          ipcMain.removeAllListeners('get-remote-module-path')
-          expect(arg).to.equal('hi child window')
-          done()
-        })
-      })
-
-      it('validates process APIs access in sandboxed renderer', (done) => {
-        ipcMain.once('answer', function (event, test) {
-          expect(test.hasCrash).to.be.true()
-          expect(test.hasHang).to.be.true()
-          expect(test.heapStatistics).to.be.an('object')
-          expect(test.blinkMemoryInfo).to.be.an('object')
-          expect(test.processMemoryInfo).to.be.an('object')
-          expect(test.systemVersion).to.be.a('string')
-          expect(test.cpuUsage).to.be.an('object')
-          expect(test.ioCounters).to.be.an('object')
-          expect(test.arch).to.equal(remote.process.arch)
-          expect(test.platform).to.equal(remote.process.platform)
-          expect(test.env).to.deep.equal(remote.process.env)
-          expect(test.execPath).to.equal(remote.process.helperExecPath)
-          expect(test.sandboxed).to.be.true()
-          expect(test.type).to.equal('renderer')
-          expect(test.version).to.equal(remote.process.version)
-          expect(test.versions).to.deep.equal(remote.process.versions)
-
-          if (process.platform === 'linux' && test.osSandbox) {
-            expect(test.creationTime).to.be.null()
-            expect(test.systemMemoryInfo).to.be.null()
-          } else {
-            expect(test.creationTime).to.be.a('number')
-            expect(test.systemMemoryInfo).to.be.an('object')
-          }
-
-          done()
-        })
-        remote.process.env.sandboxmain = 'foo'
-        w.destroy()
-        w = new BrowserWindow({
-          show: false,
-          webPreferences: {
-            sandbox: true,
-            preload
-          }
-        })
-        w.webContents.once('preload-error', (event, preloadPath, error) => {
-          done(error)
-        })
-        w.loadFile(path.join(fixtures, 'api', 'preload.html'))
-      })
-
-      it('webview in sandbox renderer', async () => {
-        w.destroy()
-        w = new BrowserWindow({
-          show: false,
-          webPreferences: {
-            sandbox: true,
-            preload,
-            webviewTag: true
-          }
-        })
-        const didAttachWebview = emittedOnce(w.webContents, 'did-attach-webview')
-        const webviewDomReady = emittedOnce(ipcMain, 'webview-dom-ready')
-        w.loadFile(path.join(fixtures, 'pages', 'webview-did-attach-event.html'))
-
-        const [, webContents] = await didAttachWebview
-        const [, id] = await webviewDomReady
-        expect(webContents.id).to.equal(id)
-      })
-    })
-
-    describe('nativeWindowOpen option', () => {
-      const networkSandbox = createNetworkSandbox(protocol)
-
-      beforeEach(async () => {
-        // used to create cross-origin navigation situations
-        await networkSandbox.serveFileFromProtocol('foo', path.join(fixtures, 'api', 'window-open-location-change.html'))
-        await networkSandbox.serveFileFromProtocol('bar', path.join(fixtures, 'api', 'window-open-location-final.html'))
-
-        w.destroy()
-        w = new BrowserWindow({
-          show: false,
-          webPreferences: {
-            nodeIntegration: true,
-            nativeWindowOpen: true,
-            // tests relies on preloads in opened windows
-            nodeIntegrationInSubFrames: true
-          }
-        })
-      })
-
-      afterEach(async () => {
-        await networkSandbox.reset()
-      })
-
-      it('opens window of about:blank with cross-scripting enabled', (done) => {
-        ipcMain.once('answer', (event, content) => {
-          expect(content).to.equal('Hello')
-          done()
-        })
-        w.loadFile(path.join(fixtures, 'api', 'native-window-open-blank.html'))
-      })
-      it('opens window of same domain with cross-scripting enabled', (done) => {
-        ipcMain.once('answer', (event, content) => {
-          expect(content).to.equal('Hello')
-          done()
-        })
-        w.loadFile(path.join(fixtures, 'api', 'native-window-open-file.html'))
-      })
-      it('blocks accessing cross-origin frames', (done) => {
-        ipcMain.once('answer', (event, content) => {
-          expect(content).to.equal('Blocked a frame with origin "file://" from accessing a cross-origin frame.')
-          done()
-        })
-        w.loadFile(path.join(fixtures, 'api', 'native-window-open-cross-origin.html'))
-      })
-      it('opens window from <iframe> tags', (done) => {
-        ipcMain.once('answer', (event, content) => {
-          expect(content).to.equal('Hello')
-          done()
-        })
-        w.loadFile(path.join(fixtures, 'api', 'native-window-open-iframe.html'))
-      });
-      (nativeModulesEnabled ? it : it.skip)('loads native addons correctly after reload', (done) => {
-        ipcMain.once('answer', (event, content) => {
-          expect(content).to.equal('function')
-          ipcMain.once('answer', (event, content) => {
-            expect(content).to.equal('function')
-            done()
-          })
-          w.reload()
-        })
-        w.loadFile(path.join(fixtures, 'api', 'native-window-open-native-addon.html'))
-      })
-      it('<webview> works in a scriptable popup', (done) => {
-        const preload = path.join(fixtures, 'api', 'new-window-webview-preload.js')
-
-        w.destroy()
-        w = new BrowserWindow({
-          show: false,
-          webPreferences: {
-            nodeIntegrationInSubFrames: true,
-            nativeWindowOpen: true,
-            webviewTag: true,
-            preload
-          }
-        })
-
-        ipcRenderer.send('set-options-on-next-new-window', w.webContents.id, 'show', false)
-
-        ipcMain.once('webview-loaded', () => {
-          done()
-        })
-        w.loadFile(path.join(fixtures, 'api', 'new-window-webview.html'))
-      })
-      it('should inherit the nativeWindowOpen setting in opened windows', (done) => {
-        w.destroy()
-        w = new BrowserWindow({
-          show: false,
-          webPreferences: {
-            nativeWindowOpen: true,
-            // test relies on preloads in opened window
-            nodeIntegrationInSubFrames: true
-          }
-        })
-
-        const preloadPath = path.join(fixtures, 'api', 'new-window-preload.js')
-        ipcRenderer.send('set-web-preferences-on-next-new-window', w.webContents.id, 'preload', preloadPath)
-        ipcMain.once('answer', (event, args) => {
-          expect(args).to.include('--native-window-open')
-          done()
-        })
-        w.loadFile(path.join(fixtures, 'api', 'new-window.html'))
-      })
-      it('should open windows with the options configured via new-window event listeners', (done) => {
-        w.destroy()
-        w = new BrowserWindow({
-          show: false,
-          webPreferences: {
-            nativeWindowOpen: true,
-            // test relies on preloads in opened window
-            nodeIntegrationInSubFrames: true
-          }
-        })
-
-        const preloadPath = path.join(fixtures, 'api', 'new-window-preload.js')
-        ipcRenderer.send('set-web-preferences-on-next-new-window', w.webContents.id, 'preload', preloadPath)
-        ipcRenderer.send('set-web-preferences-on-next-new-window', w.webContents.id, 'foo', 'bar')
-        ipcMain.once('answer', (event, args, webPreferences) => {
-          expect(webPreferences.foo).to.equal('bar')
-          done()
-        })
-        w.loadFile(path.join(fixtures, 'api', 'new-window.html'))
-      })
-      it('retains the original web preferences when window.location is changed to a new origin', async () => {
-        w.destroy()
-        w = new BrowserWindow({
-          show: true,
-          webPreferences: {
-            nativeWindowOpen: true,
-            // test relies on preloads in opened window
-            nodeIntegrationInSubFrames: true
-          }
-        })
-
-        ipcRenderer.send('set-web-preferences-on-next-new-window', w.webContents.id, 'preload', path.join(fixtures, 'api', 'window-open-preload.js'))
-        const p = emittedOnce(ipcMain, 'answer')
-        w.loadFile(path.join(fixtures, 'api', 'window-open-location-open.html'))
-        const [, args, typeofProcess] = await p
-        expect(args).not.to.include('--node-integration')
-        expect(args).to.include('--native-window-open')
-        expect(typeofProcess).to.eql('undefined')
-      })
-
-      it('window.opener is not null when window.location is changed to a new origin', async () => {
-        w.destroy()
-        w = new BrowserWindow({
-          show: true,
-          webPreferences: {
-            nativeWindowOpen: true,
-            // test relies on preloads in opened window
-            nodeIntegrationInSubFrames: true
-          }
-        })
-
-        ipcRenderer.send('set-web-preferences-on-next-new-window', w.webContents.id, 'preload', path.join(fixtures, 'api', 'window-open-preload.js'))
-        const p = emittedOnce(ipcMain, 'answer')
-        w.loadFile(path.join(fixtures, 'api', 'window-open-location-open.html'))
-        const [, , , windowOpenerIsNull] = await p
-        expect(windowOpenerIsNull).to.be.false('window.opener is null')
-      })
-
-      it('should have nodeIntegration disabled in child windows', async () => {
-        w.destroy()
-        w = new BrowserWindow({
-          show: false,
-          webPreferences: {
-            nodeIntegration: true,
-            nativeWindowOpen: true
-          }
-        })
-        const p = emittedOnce(ipcMain, 'answer')
-        w.loadFile(path.join(fixtures, 'api', 'native-window-open-argv.html'))
-        const [, typeofProcess] = await p
-        expect(typeofProcess).to.eql('undefined')
-      })
-    })
-
-    describe('"disableHtmlFullscreenWindowResize" option', () => {
-      it('prevents window from resizing when set', (done) => {
-        w.destroy()
-        w = new BrowserWindow({
-          show: false,
-          webPreferences: {
-            disableHtmlFullscreenWindowResize: true
-          }
-        })
-        w.webContents.once('did-finish-load', () => {
-          const size = w.getSize()
-          w.webContents.once('enter-html-full-screen', () => {
-            const newSize = w.getSize()
-            expect(newSize).to.deep.equal(size)
-            done()
-          })
-          w.webContents.executeJavaScript('document.body.webkitRequestFullscreen()', true)
-        })
-        w.loadURL('about:blank')
-      })
-    })
-  })
-
-  describe('nativeWindowOpen + contextIsolation options', () => {
-    beforeEach(() => {
-      w.destroy()
-      w = new BrowserWindow({
-        show: false,
-        webPreferences: {
-          nativeWindowOpen: true,
-          contextIsolation: true,
-          preload: path.join(fixtures, 'api', 'native-window-open-isolated-preload.js')
-        }
-      })
-    })
-
-    it('opens window with cross-scripting enabled from isolated context', (done) => {
-      ipcMain.once('answer', (event, content) => {
-        expect(content).to.equal('Hello')
-        done()
-      })
-      w.loadFile(path.join(fixtures, 'api', 'native-window-open-isolated.html'))
-    })
-  })
-
-  describe('beforeunload handler', () => {
-    it('returning undefined would not prevent close', (done) => {
-      w.once('closed', () => { done() })
-      w.loadFile(path.join(fixtures, 'api', 'close-beforeunload-undefined.html'))
-    })
-    it('returning false would prevent close', (done) => {
-      w.once('onbeforeunload', () => { done() })
-      w.loadFile(path.join(fixtures, 'api', 'close-beforeunload-false.html'))
-    })
-    it('returning empty string would prevent close', (done) => {
-      w.once('onbeforeunload', () => { done() })
-      w.loadFile(path.join(fixtures, 'api', 'close-beforeunload-empty-string.html'))
-    })
-    it('emits for each close attempt', (done) => {
-      let beforeUnloadCount = 0
-      w.on('onbeforeunload', () => {
-        beforeUnloadCount += 1
-        if (beforeUnloadCount < 3) {
-          w.close()
-        } else if (beforeUnloadCount === 3) {
-          done()
-        }
-      })
-      w.webContents.once('did-finish-load', () => { w.close() })
-      w.loadFile(path.join(fixtures, 'api', 'beforeunload-false-prevent3.html'))
-    })
-    it('emits for each reload attempt', (done) => {
-      let beforeUnloadCount = 0
-      w.on('onbeforeunload', () => {
-        beforeUnloadCount += 1
-        if (beforeUnloadCount < 3) {
-          w.reload()
-        } else if (beforeUnloadCount === 3) {
-          done()
-        }
-      })
-      w.webContents.once('did-finish-load', () => {
-        w.webContents.once('did-finish-load', () => {
-          expect.fail('Reload was not prevented')
-        })
-        w.reload()
-      })
-      w.loadFile(path.join(fixtures, 'api', 'beforeunload-false-prevent3.html'))
-    })
-    it('emits for each navigation attempt', (done) => {
-      let beforeUnloadCount = 0
-      w.on('onbeforeunload', () => {
-        beforeUnloadCount += 1
-        if (beforeUnloadCount < 3) {
-          w.loadURL('about:blank')
-        } else if (beforeUnloadCount === 3) {
-          done()
-        }
-      })
-      w.webContents.once('did-finish-load', () => {
-        w.webContents.once('did-finish-load', () => {
-          expect.fail('Navigation was not prevented')
-        })
-        w.loadURL('about:blank')
-      })
-      w.loadFile(path.join(fixtures, 'api', 'beforeunload-false-prevent3.html'))
-    })
-  })
-
-  // visibilitychange event is broken upstream, see crbug.com/920839
-  xdescribe('document.visibilityState/hidden', () => {
-    beforeEach(() => { w.destroy() })
-
-    function onVisibilityChange (callback) {
-      ipcMain.on('pong', (event, visibilityState, hidden) => {
-        if (event.sender.id === w.webContents.id) {
-          callback(visibilityState, hidden)
-        }
-      })
-    }
-
-    function onNextVisibilityChange (callback) {
-      ipcMain.once('pong', (event, visibilityState, hidden) => {
-        if (event.sender.id === w.webContents.id) {
-          callback(visibilityState, hidden)
-        }
-      })
-    }
-
-    afterEach(() => { ipcMain.removeAllListeners('pong') })
-
-    it('visibilityState is initially visible despite window being hidden', (done) => {
-      w = new BrowserWindow({
-        show: false,
-        width: 100,
-        height: 100,
-        webPreferences: {
-          nodeIntegration: true
-        }
-      })
-
-      let readyToShow = false
-      w.once('ready-to-show', () => {
-        readyToShow = true
-      })
-
-      onNextVisibilityChange((visibilityState, hidden) => {
-        expect(readyToShow).to.be.false()
-        expect(visibilityState).to.equal('visible')
-        expect(hidden).to.be.false()
-
-        done()
-      })
-
-      w.loadFile(path.join(fixtures, 'pages', 'visibilitychange.html'))
-    })
-    it('visibilityState changes when window is hidden', (done) => {
-      w = new BrowserWindow({
-        width: 100,
-        height: 100,
-        webPreferences: {
-          nodeIntegration: true
-        }
-      })
-
-      onNextVisibilityChange((visibilityState, hidden) => {
-        expect(visibilityState).to.equal('visible')
-        expect(hidden).to.be.false()
-
-        onNextVisibilityChange((visibilityState, hidden) => {
-          expect(visibilityState).to.equal('hidden')
-          expect(hidden).to.be.true()
-          done()
-        })
-
-        w.hide()
-      })
-
-      w.loadFile(path.join(fixtures, 'pages', 'visibilitychange.html'))
-    })
-    it('visibilityState changes when window is shown', (done) => {
-      w = new BrowserWindow({
-        width: 100,
-        height: 100,
-        webPreferences: {
-          nodeIntegration: true
-        }
-      })
-
-      onNextVisibilityChange((visibilityState, hidden) => {
-        onVisibilityChange((visibilityState, hidden) => {
-          if (!hidden) {
-            expect(visibilityState).to.equal('visible')
-            done()
-          }
-        })
-
-        w.hide()
-        w.show()
-      })
-
-      w.loadFile(path.join(fixtures, 'pages', 'visibilitychange.html'))
-    })
-    it('visibilityState changes when window is shown inactive', function (done) {
-      if (isCI && process.platform === 'win32') {
-        // FIXME(alexeykuzmin): Skip the test instead of marking it as passed.
-        // afterEach hook won't be run if a test is skipped dynamically.
-        // If afterEach isn't run current window won't be destroyed
-        // and the next test will fail on assertion in `closeWindow()`.
-        // this.skip()
-        return done()
-      }
-
-      w = new BrowserWindow({
-        width: 100,
-        height: 100,
-        webPreferences: {
-          nodeIntegration: true
-        }
-      })
-
-      onNextVisibilityChange((visibilityState, hidden) => {
-        onVisibilityChange((visibilityState, hidden) => {
-          if (!hidden) {
-            expect(visibilityState).to.equal('visible')
-            done()
-          }
-        })
-
-        w.hide()
-        w.showInactive()
-      })
-
-      w.loadFile(path.join(fixtures, 'pages', 'visibilitychange.html'))
-    })
-    it('visibilityState changes when window is minimized', function (done) {
-      if (isCI && process.platform === 'linux') {
-        // FIXME(alexeykuzmin): Skip the test instead of marking it as passed.
-        // afterEach hook won't be run if a test is skipped dynamically.
-        // If afterEach isn't run current window won't be destroyed
-        // and the next test will fail on assertion in `closeWindow()`.
-        // this.skip()
-        return done()
-      }
-
-      w = new BrowserWindow({
-        width: 100,
-        height: 100,
-        webPreferences: {
-          nodeIntegration: true
-        }
-      })
-
-      onNextVisibilityChange((visibilityState, hidden) => {
-        expect(visibilityState).to.equal('visible')
-        expect(hidden).to.be.false()
-
-        onNextVisibilityChange((visibilityState, hidden) => {
-          expect(visibilityState).to.equal('hidden')
-          expect(hidden).to.be.true()
-          done()
-        })
-
-        w.minimize()
-      })
-
-      w.loadFile(path.join(fixtures, 'pages', 'visibilitychange.html'))
-    })
-    it('visibilityState remains visible if backgroundThrottling is disabled', (done) => {
-      w = new BrowserWindow({
-        show: false,
-        width: 100,
-        height: 100,
-        webPreferences: {
-          backgroundThrottling: false,
-          nodeIntegration: true
-        }
-      })
-
-      onNextVisibilityChange((visibilityState, hidden) => {
-        expect(visibilityState).to.equal('visible')
-        expect(hidden).to.be.false()
-
-        onNextVisibilityChange((visibilityState, hidden) => {
-          done(new Error(`Unexpected visibility change event. visibilityState: ${visibilityState} hidden: ${hidden}`))
-        })
-      })
-
-      w.once('show', () => {
-        w.once('hide', () => {
-          w.once('show', () => {
-            done()
-          })
-          w.show()
-        })
-        w.hide()
-      })
-      w.show()
-
-      w.loadFile(path.join(fixtures, 'pages', 'visibilitychange.html'))
-    })
-  })
-
-  describe('new-window event', () => {
-    before(function () {
-      if (isCI && process.platform === 'darwin') {
-        this.skip()
-      }
-    })
-
-    it('emits when window.open is called', (done) => {
-      w.webContents.once('new-window', (e, url, frameName, disposition, options, additionalFeatures) => {
-        e.preventDefault()
-        expect(url).to.equal('http://host/')
-        expect(frameName).to.equal('host')
-        expect(additionalFeatures[0]).to.equal('this-is-not-a-standard-feature')
-        done()
-      })
-      w.loadFile(path.join(fixtures, 'pages', 'window-open.html'))
-    })
-    it('emits when window.open is called with no webPreferences', (done) => {
-      w.destroy()
-      w = new BrowserWindow({ show: false })
-      w.webContents.once('new-window', function (e, url, frameName, disposition, options, additionalFeatures) {
-        e.preventDefault()
-        expect(url).to.equal('http://host/')
-        expect(frameName).to.equal('host')
-        expect(additionalFeatures[0]).to.equal('this-is-not-a-standard-feature')
-        done()
-      })
-      w.loadFile(path.join(fixtures, 'pages', 'window-open.html'))
-    })
-
-    it('emits when link with target is called', (done) => {
-      w.webContents.once('new-window', (e, url, frameName) => {
-        e.preventDefault()
-        expect(url).to.equal('http://host/')
-        expect(frameName).to.equal('target')
-        done()
-      })
-      w.loadFile(path.join(fixtures, 'pages', 'target-name.html'))
-    })
-  })
-
-  describe('maximize event', () => {
-    if (isCI) return
-
-    it('emits when window is maximized', (done) => {
-      w.once('maximize', () => { done() })
-      w.show()
-      w.maximize()
-    })
-  })
-
-  describe('unmaximize event', () => {
-    if (isCI) return
-
-    it('emits when window is unmaximized', (done) => {
-      w.once('unmaximize', () => { done() })
-      w.show()
-      w.maximize()
-      w.unmaximize()
-    })
-  })
-
-  describe('minimize event', () => {
-    if (isCI) return
-
-    it('emits when window is minimized', (done) => {
-      w.once('minimize', () => { done() })
-      w.show()
-      w.minimize()
-    })
-  })
-
-  describe('sheet-begin event', () => {
-    let sheet = null
-
-    before(function () {
-      if (process.platform !== 'darwin') {
-        this.skip()
-      }
-    })
-
-    afterEach(() => {
-      return closeWindow(sheet, { assertSingleWindow: false }).then(() => { sheet = null })
-    })
-
-    it('emits when window opens a sheet', (done) => {
-      w.show()
-      w.once('sheet-begin', () => {
-        sheet.close()
-        done()
-      })
-      sheet = new BrowserWindow({
-        modal: true,
-        parent: w
-      })
-    })
-  })
-
-  describe('sheet-end event', () => {
-    let sheet = null
-
-    before(function () {
-      if (process.platform !== 'darwin') {
-        this.skip()
-      }
-    })
-
-    afterEach(() => {
-      return closeWindow(sheet, { assertSingleWindow: false }).then(() => { sheet = null })
-    })
-
-    it('emits when window has closed a sheet', (done) => {
-      w.show()
-      sheet = new BrowserWindow({
-        modal: true,
-        parent: w
-      })
-      w.once('sheet-end', () => { done() })
-      sheet.close()
-    })
-  })
-
-  describe('beginFrameSubscription method', () => {
-    it('does not crash when callback returns nothing', (done) => {
-      w.loadFile(path.join(fixtures, 'api', 'frame-subscriber.html'))
-      w.webContents.on('dom-ready', () => {
-        w.webContents.beginFrameSubscription(function (data) {
-          // Pending endFrameSubscription to next tick can reliably reproduce
-          // a crash which happens when nothing is returned in the callback.
-          setTimeout(() => {
-            w.webContents.endFrameSubscription()
-            done()
-          })
-        })
-      })
-    })
-
-    it('subscribes to frame updates', (done) => {
-      let called = false
-      w.loadFile(path.join(fixtures, 'api', 'frame-subscriber.html'))
-      w.webContents.on('dom-ready', () => {
-        w.webContents.beginFrameSubscription(function (data) {
-          // This callback might be called twice.
-          if (called) return
-          called = true
-
-          expect(data.constructor.name).to.equal('NativeImage')
-          expect(data.isEmpty()).to.be.false()
-
-          w.webContents.endFrameSubscription()
-          done()
-        })
-      })
-    })
-
-    it('subscribes to frame updates (only dirty rectangle)', (done) => {
-      let called = false
-      let gotInitialFullSizeFrame = false
-      const [contentWidth, contentHeight] = w.getContentSize()
-      w.webContents.on('did-finish-load', () => {
-        w.webContents.beginFrameSubscription(true, (image, rect) => {
-          if (image.isEmpty()) {
-            // Chromium sometimes sends a 0x0 frame at the beginning of the
-            // page load.
-            return
-          }
-          if (rect.height === contentHeight && rect.width === contentWidth &&
-              !gotInitialFullSizeFrame) {
-            // The initial frame is full-size, but we're looking for a call
-            // with just the dirty-rect. The next frame should be a smaller
-            // rect.
-            gotInitialFullSizeFrame = true
-            return
-          }
-          // This callback might be called twice.
-          if (called) return
-          // We asked for just the dirty rectangle, so we expect to receive a
-          // rect smaller than the full size.
-          // TODO(jeremy): this is failing on windows currently; investigate.
-          // assert(rect.width < contentWidth || rect.height < contentHeight)
-          called = true
-
-          const expectedSize = rect.width * rect.height * 4
-          expect(image.getBitmap()).to.be.an.instanceOf(Buffer).with.lengthOf(expectedSize)
-          w.webContents.endFrameSubscription()
-          done()
-        })
-      })
-      w.loadFile(path.join(fixtures, 'api', 'frame-subscriber.html'))
-    })
-
-    it('throws error when subscriber is not well defined', (done) => {
-      w.loadFile(path.join(fixtures, 'api', 'frame-subscriber.html'))
-      try {
-        w.webContents.beginFrameSubscription(true, true)
-      } catch (e) {
-        done()
-      }
-    })
-  })
-
-  describe('savePage method', () => {
-    const savePageDir = path.join(fixtures, 'save_page')
-    const savePageHtmlPath = path.join(savePageDir, 'save_page.html')
-    const savePageJsPath = path.join(savePageDir, 'save_page_files', 'test.js')
-    const savePageCssPath = path.join(savePageDir, 'save_page_files', 'test.css')
-
-    after(() => {
-      try {
-        fs.unlinkSync(savePageCssPath)
-        fs.unlinkSync(savePageJsPath)
-        fs.unlinkSync(savePageHtmlPath)
-        fs.rmdirSync(path.join(savePageDir, 'save_page_files'))
-        fs.rmdirSync(savePageDir)
-      } catch (e) {
-        // Ignore error
-      }
-    })
-
-    it('should save page to disk', async () => {
-      await w.loadFile(path.join(fixtures, 'pages', 'save_page', 'index.html'))
-      await w.webContents.savePage(savePageHtmlPath, 'HTMLComplete')
-
-      expect(fs.existsSync(savePageHtmlPath)).to.be.true()
-      expect(fs.existsSync(savePageJsPath)).to.be.true()
-      expect(fs.existsSync(savePageCssPath)).to.be.true()
-    })
-  })
-
-  describe('BrowserWindow options argument is optional', () => {
-    it('should create a window with default size (800x600)', () => {
-      w.destroy()
-      w = new BrowserWindow()
-      const size = w.getSize()
-      expect(size).to.deep.equal([800, 600])
-    })
-  })
 
   describe('window states', () => {
     it('does not resize frameless windows when states change', () => {
@@ -1968,207 +589,6 @@ describe('BrowserWindow module', () => {
         expect(w.hasShadow()).to.be.false()
         w.setHasShadow(true)
         expect(w.hasShadow()).to.be.true()
-      })
-    })
-  })
-
-  describe('BrowserWindow.restore()', () => {
-    it('should restore the previous window size', () => {
-      if (w != null) w.destroy()
-
-      w = new BrowserWindow({
-        minWidth: 800,
-        width: 800
-      })
-
-      const initialSize = w.getSize()
-      w.minimize()
-      w.restore()
-      expectBoundsEqual(w.getSize(), initialSize)
-    })
-  })
-
-  describe('BrowserWindow.unmaximize()', () => {
-    it('should restore the previous window position', () => {
-      if (w != null) w.destroy()
-      w = new BrowserWindow()
-
-      const initialPosition = w.getPosition()
-      w.maximize()
-      w.unmaximize()
-      expectBoundsEqual(w.getPosition(), initialPosition)
-    })
-  })
-
-  describe('BrowserWindow.setFullScreen(false)', () => {
-    before(function () {
-      // only applicable to windows: https://github.com/electron/electron/issues/6036
-      if (process.platform !== 'win32') {
-        this.skip()
-      }
-    })
-
-    it('should restore a normal visible window from a fullscreen startup state', (done) => {
-      w.webContents.once('did-finish-load', () => {
-        // start fullscreen and hidden
-        w.setFullScreen(true)
-        w.once('show', () => { w.setFullScreen(false) })
-        w.once('leave-full-screen', () => {
-          expect(w.isVisible()).to.be.true()
-          expect(w.isFullScreen()).to.be.false()
-          done()
-        })
-        w.show()
-      })
-      w.loadURL('about:blank')
-    })
-    it('should keep window hidden if already in hidden state', (done) => {
-      w.webContents.once('did-finish-load', () => {
-        w.once('leave-full-screen', () => {
-          expect(w.isVisible()).to.be.false()
-          expect(w.isFullScreen()).to.be.false()
-          done()
-        })
-        w.setFullScreen(false)
-      })
-      w.loadURL('about:blank')
-    })
-  })
-
-  describe('BrowserWindow.setFullScreen(false) when HTML fullscreen', () => {
-    before(function () {
-      if (process.platform !== 'darwin') {
-        this.skip()
-      }
-    })
-
-    it('exits HTML fullscreen when window leaves fullscreen', (done) => {
-      w.destroy()
-      w = new BrowserWindow()
-      w.webContents.once('did-finish-load', () => {
-        w.webContents.executeJavaScript('document.body.webkitRequestFullscreen()', true).then(() => {
-          w.once('enter-full-screen', () => {
-            w.once('leave-html-full-screen', () => {
-              done()
-            })
-            w.setFullScreen(false)
-          })
-        })
-      })
-      w.loadURL('about:blank')
-    })
-  })
-
-  describe('parent window', () => {
-    let c = null
-
-    beforeEach(() => {
-      if (c != null) c.destroy()
-      c = new BrowserWindow({ show: false, parent: w })
-    })
-
-    afterEach(() => {
-      if (c != null) c.destroy()
-      c = null
-    })
-
-    describe('parent option', () => {
-      it('sets parent window', () => {
-        expect(c.getParentWindow()).to.equal(w)
-      })
-      it('adds window to child windows of parent', () => {
-        expect(w.getChildWindows()).to.deep.equal([c])
-      })
-      it('removes from child windows of parent when window is closed', (done) => {
-        c.once('closed', () => {
-          expect(w.getChildWindows()).to.deep.equal([])
-          done()
-        })
-        c.close()
-      })
-
-      it('should not affect the show option', () => {
-        expect(c.isVisible()).to.be.false()
-        expect(c.getParentWindow().isVisible()).to.be.false()
-      })
-    })
-
-    describe('win.setParentWindow(parent)', () => {
-      beforeEach(() => {
-        if (c != null) c.destroy()
-        c = new BrowserWindow({ show: false })
-      })
-
-      it('sets parent window', () => {
-        expect(w.getParentWindow()).to.be.null()
-        expect(c.getParentWindow()).to.be.null()
-        c.setParentWindow(w)
-        expect(c.getParentWindow()).to.equal(w)
-        c.setParentWindow(null)
-        expect(c.getParentWindow()).to.be.null()
-      })
-      it('adds window to child windows of parent', () => {
-        expect(w.getChildWindows()).to.deep.equal([])
-        c.setParentWindow(w)
-        expect(w.getChildWindows()).to.deep.equal([c])
-        c.setParentWindow(null)
-        expect(w.getChildWindows()).to.deep.equal([])
-      })
-      it('removes from child windows of parent when window is closed', (done) => {
-        c.once('closed', () => {
-          expect(w.getChildWindows()).to.deep.equal([])
-          done()
-        })
-        c.setParentWindow(w)
-        c.close()
-      })
-    })
-
-    describe('modal option', () => {
-      before(function () {
-        // The isEnabled API is not reliable on macOS.
-        if (process.platform === 'darwin') {
-          this.skip()
-        }
-      })
-
-      beforeEach(() => {
-        if (c != null) c.destroy()
-        c = new BrowserWindow({ show: false, parent: w, modal: true })
-      })
-
-      it('disables parent window', () => {
-        expect(w.isEnabled()).to.be.true()
-        c.show()
-        expect(w.isEnabled()).to.be.false()
-      })
-      it('re-enables an enabled parent window when closed', (done) => {
-        c.once('closed', () => {
-          expect(w.isEnabled()).to.be.true()
-          done()
-        })
-        c.show()
-        c.close()
-      })
-      it('does not re-enable a disabled parent window when closed', (done) => {
-        c.once('closed', () => {
-          expect(w.isEnabled()).to.be.false()
-          done()
-        })
-        w.setEnabled(false)
-        c.show()
-        c.close()
-      })
-      it('disables parent window recursively', () => {
-        const c2 = new BrowserWindow({ show: false, parent: w, modal: true })
-        c.show()
-        expect(w.isEnabled()).to.be.false()
-        c2.show()
-        expect(w.isEnabled()).to.be.false()
-        c.destroy()
-        expect(w.isEnabled()).to.be.false()
-        c2.destroy()
-        expect(w.isEnabled()).to.be.true()
       })
     })
   })
